requirements.txt
django==5.1.3
psycopg2-binary==2.9.10
python-telegram-bot==21.7
python-dotenv==1.0.1
djangorestframework==3.15.2